+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Do not show "Due" date label in canvas assignments if there is no value [#3013](https://github.com/rokwire/illinois-app/issues/3013).
+
+## [4.3.20] - 2023-03-02
+### Added
+- Bring back old Canvas UI with debug switch [#3011](https://github.com/rokwire/illinois-app/issues/3011).
 
 ## [4.3.19] - 2023-02-23
 ### Added